Handle error and forwarding classes from Swift to Obj-C

+++ Problem Set/Guitar_SwiftSpecific/Guitar_SwiftSpecific/GuitarString.swift
@@ -9,7 +9,7 @@
 import Cocoa
 
 
-enum Error: Int, ErrorType {
+enum Error: NSInteger, ErrorType {
     case Broken = 997
     case OutOfTune = 998
 }
