+++ Problem Set/RegiftErrorEnum.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 // Errors thrown by Regift
-@obj public enum RegiftError: Int, ErrorType {
+@obj public enum RegiftError: NSInteger, ErrorType {
     case DestinationNotFound = 997
     case AddFrameToDestination = 998
     case DestinationFinalize = 999
